code/requirements.txt
streamlit
pandas
numpy
matplotlib
seaborn
scikit-learn
plotly
pillow
tensorflow
keras
tqdm
faiss-cpu
